Вынес алерты в модуль
